Conflicts:
	src/java/com/android/internal/telephony/InboundSmsHandler.java
	src/java/com/android/internal/telephony/ModemBindingPolicyHandler.java
	src/java/com/android/internal/telephony/PhoneBase.java
	src/java/com/android/internal/telephony/PhoneFactory.java
	src/java/com/android/internal/telephony/SubInfoRecordUpdater.java
	src/java/com/android/internal/telephony/SubscriptionController.java
	src/java/com/android/internal/telephony/cdma/CDMAPhone.java
	src/java/com/android/internal/telephony/gsm/GSMPhone.java
	src/java/com/android/internal/telephony/imsphone/ImsPhoneCallTracker.java
	src/java/com/android/internal/telephony/util/BlacklistUtils.java

+++ src/java/com/android/internal/telephony/dataconnection/DataResetEventTracker.java
@@ -122,7 +122,7 @@ public class DataResetEventTracker {
             mPreviousLocation = (GsmCellLocation) mPhone.getCellLocation();
             if (DBG) log("DataConnection mPreviousLocation : " + mPreviousLocation);
         }
-        long subId = SubscriptionManager.getDefaultDataSubId();
+        int subId = SubscriptionManager.getDefaultDataSubId();
 
         if (mPhoneStateListener == null) {
             mPhoneStateListener = new PhoneStateListener(subId) {
